+++ lib/formotion/row_type/button.rb
@@ -23,13 +23,6 @@ module Formotion
         end
         nil
       end
-
-      def on_select(tableView, tableViewDelegate)
-        # Override in subclasses
-        if self.row.on_tap_callback
-          self.row.on_tap_callback.call(self.row)
-        end
-      end
     end
   end
 end